+++ src/uk/co/quarklike/prototype/map/entity/EntityProjectileItem.java
@@ -10,12 +10,12 @@ public class EntityProjectileItem extends EntityProjectile {
 	public EntityProjectileItem(byte direction, int itemID) {
 		super(Item.getItem(itemID).getName(), Item.getItem(itemID).getTexture(), direction, Physics.maxRange(15, 1f, 1.5f));
 		this.itemID = itemID;
-		this.body.addItem(itemID, 1);
+		this.body.addItem(itemID, (byte) 1);
 	}
 
 	@Override
 	public void finish() {
-		this.dropItem(new ItemStack(this.itemID, 1));
+		this.dropItem(new ItemStack(this.itemID, (byte) 1));
 		super.finish();
 	}
